+++ benches/dowser.rs
@@ -10,30 +10,29 @@ use dowser::{
 	Dowser,
 	Extension,
 };
-use std::time::Duration;
 
 
 const GZ: Extension = Extension::new2(*b"gz");
 
 
 benches!(
-	Bench::new("dowser::Dowser", "from(/usr/share)")
-		.timed(Duration::from_secs(6))
-		.with(|| Dowser::from("/usr/share").collect::<Vec<_>>()),
+	Bench::new("dowser::Dowser::from(/usr/share).collect()")
+		.run(|| Dowser::from("/usr/share").collect::<Vec<_>>()),
+
+	Bench::new("dowser::Dowser::from(/usr/share).into_vec()")
+		.run(|| Dowser::from("/usr/share").into_vec()),
 
 	Bench::spacer(),
 
-	Bench::new("dowser::Dowser::from(/usr/share)", "filter(gz)")
-		.timed(Duration::from_secs(6))
-		.with_setup(
+	Bench::new("dowser::Dowser::from(/usr/share).filter(gz).collect()")
+		.run_seeded(
 			Dowser::from("/usr/share"),
 			|d| d.filter(|p| Some(GZ) == Extension::try_from2(p)).collect::<Vec<_>>()
 		),
 
-	Bench::new("dowser::Dowser::from(/usr/share)", "into_vec(gz)")
-		.timed(Duration::from_secs(6))
-		.with_setup(
+	Bench::new("dowser::Dowser::from(/usr/share).into_vec_filtered(gz)")
+		.run_seeded(
 			Dowser::from("/usr/share"),
-			|d| d.into_vec(|p| Some(GZ) == Extension::try_from2(p))
+			|d| d.into_vec_filtered(|p| Some(GZ) == Extension::try_from2(p))
 		),
 );