+++ examples/dowser.rs
@@ -2,24 +2,25 @@
 # Dowser: Filtered Find
 */
 
-use dowser::{
-	Dowser,
-	Extension,
-};
-use std::{
-	path::PathBuf,
-	time::Instant,
-};
-
-/// Do it.
+#[cfg(unix)]
+/// # Do it.
 fn main() {
+	use dowser::{
+		Dowser,
+		Extension,
+	};
+	use std::{
+		path::PathBuf,
+		time::Instant,
+	};
+
 	const EXT: Extension = Extension::new2(*b"gz");
 
 	// Search for gzipped MAN pages.
 	let now = Instant::now();
 	let files: Vec<PathBuf> = Dowser::default()
 		.with_path("/usr/share")
-		.into_vec(|p| Some(EXT) == Extension::try_from2(p));
+		.into_vec_filtered(|p| Some(EXT) == Extension::try_from2(p));
 
 	println!("Search took {} seconds.", now.elapsed().as_millis() as f64 / 1000.000);
 
@@ -30,4 +31,10 @@ fn main() {
 	else {
 		println!("There are {} .gz files in /usr/share/.", files.len());
 	}
+}
+
+#[cfg(not(unix))]
+/// # Don't Do It.
+fn main() {
+	println!("This example is only for unix.");
 }