+++ src/lib.rs
@@ -1,7 +1,12 @@
 /*!
 # Dowser
 
-[`Dowser`] is a(nother) fast, multi-threaded, recursive file-finding library for Unix/Rust. It differs from [`Walkdir`](https://crates.io/crates/walkdir) and kin in a number of ways:
+[![Documentation](https://docs.rs/dowser/badge.svg)](https://docs.rs/dowser/)
+[![Changelog](https://img.shields.io/crates/v/dowser.svg?label=Changelog&color=9cf)](https://github.com/Blobfolio/dowser/blob/master/CHANGELOG.md)
+[![crates.io](https://img.shields.io/crates/v/dowser.svg)](https://crates.io/crates/dowser)
+[![PRs Welcome](https://img.shields.io/badge/PRs-welcome-brightgreen.svg?style=flat-square)](https://github.com/Blobfolio/dowser)
+
+[`Dowser`] is a(nother) fast, recursive file-finding library for Unix/Rust. It differs from [`Walkdir`](https://crates.io/crates/walkdir) and kin in a number of ways:
 
 * It is not limited to one root; any number of file and directory paths can be loaded and traversed en masse;
 * Symlinks and hidden directories are followed like any other, including across devices;
@@ -9,11 +14,7 @@
 
 If those things sound nice, this library might be a good fit.
 
-On the other hand, [`Dowser`] is optimized for _file_ searching; the iterator crawls but does not yield directory paths.
-
-Additionally, path deduping relies on Unix metadata; **this library is not compatible with Windows**.
-
-Depending on your needs, those limitations could be bad, in which case something like [`Walkdir`](https://crates.io/crates/walkdir) would make more sense.
+On the other hand, [`Dowser`] is optimized for _file_ searching; the iterator crawls but does not yield directory paths, which could be bad if you need those too. Haha.
 
 
 
@@ -23,51 +24,58 @@ Add `dowser` to your `dependencies` in `Cargo.toml`, like:
 
 ```text,ignore
 [dependencies]
-dowser = "0.5.*"
+dowser = "0.6.*"
 ```
 
 
 
 ## Example
 
-All you need to do is chain [`Dowser::default`] with one or more of the
-following seed methods:
+All you need to do is chain [`Dowser::default`] with one or more of the following seed methods:
 
 * [`Dowser::with_path`] / [`Dowser::with_paths`]
 * [`Dowser::without_path`] / [`Dowser::without_paths`]
 
-From there, you can use whatever `Iterator` methods you want.
+From there, you can apply any [`Iterator`](std::iter::Iterator) methods you want, or immediately collect the results using [`Dowser::into_vec`] or [`Dowser::into_vec_filtered`].
 
 ```
-use dowser::{
-    Dowser,
-    Extension,
-};
-use std::os::unix::ffi::OsStrExt;
+use dowser::Dowser;
 use std::path::PathBuf;
 
 // Return all files under "/usr/share/man".
-let files: Vec::<PathBuf> = Dowser::default()
+let files1: Vec::<PathBuf> = Dowser::default()
     .with_path("/usr/share/man")
     .collect();
 
+// Same as above, but slightly faster.
+let files2: Vec::<PathBuf> = Dowser::default()
+    .with_path("/usr/share/man")
+    .into_vec();
+
+assert_eq!(files1.len(), files2.len());
+
 // Return only Gzipped files using callback filter.
-let files: Vec::<PathBuf> = Dowser::default()
+let files1: Vec::<PathBuf> = Dowser::default()
     .with_path("/usr/share/man")
     .filter(|p|
         p.extension().map_or(
             false,
-            |e| e.as_bytes().eq_ignore_ascii_case(b"gz")
+            |e| e.eq_ignore_ascii_case("gz")
         )
     )
     .collect();
 
-// Same Gzip example, but using Extension.
-const EXT: Extension = Extension::new2(*b"gz");
-let files: Vec::<PathBuf> = Dowser::default()
+// Same as above, but slightly faster.
+let files2: Vec::<PathBuf> = Dowser::default()
     .with_path("/usr/share/man")
-    .filter(|p| Some(EXT) == Extension::try_from2(p))
-    .collect();
+    .into_vec_filtered(|p|
+        p.extension().map_or(
+            false,
+            |e| e.eq_ignore_ascii_case("gz")
+        )
+    );
+
+assert_eq!(files1.len(), files2.len());
 ```
 */
 
@@ -95,17 +103,27 @@ let files: Vec::<PathBuf> = Dowser::default()
 	unused_extern_crates,
 	unused_import_braces,
 )]
+
 #![allow(clippy::module_name_repetitions)] // This is fine.
 
+#![cfg_attr(feature = "docsrs", feature(doc_cfg))]
 
 
-mod ext;
+
 mod entry;
 mod iter;
+
+#[cfg(unix)] mod ext;
+
+#[cfg(unix)]
+#[cfg_attr(feature = "docsrs", doc(cfg(unix)))]
 pub mod utility;
 
 
 
 pub(crate) use entry::Entry;
-pub use ext::Extension;
 pub use iter::Dowser;
+
+#[cfg(unix)]
+#[cfg_attr(feature = "docsrs", doc(cfg(unix)))]
+pub use ext::Extension;