+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 
 
+## [0.6.0](https://github.com/Blobfolio/dowser/releases/tag/v0.6.0) - TBD
+
+### New
+
+* `Dowser::into_vec_filtered` (_f._ `Dowser::into_vec`)
+* `Dowser::into_vec`
+
+### Changed
+
+* Traversal is now fully serial. (Multi-threading came with too many gotchas and didn't improve performance much for most workloads.)
+* `Dowser::with_paths` and `Dowser::without_paths` — the plural methods — will now explicitly panic if passed a direct `Path` or `PathBuf` instead of a _proper_ `Iterator` of paths.
+* `Dowser` now has basic Windows support.
+
+
+
 ## [0.5.3](https://github.com/Blobfolio/dowser/releases/tag/v0.5.3) - 2022-08-11
 
 ### Changed